requirements.txt
requests
mysql-connector-python>=8.0.32, <8.1.0
python-dotenv